+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add shrink-on-load using the "page" param to kakaduload
 - add "options" flag to kakadusave
 - add resolution load and save
+- add "htj2k" arg to enable high-throughput j2k compression
 
 ## 2024/1/11 1.0